Redux: reorganized
