+++ frontend/src/redux/store.ts
@@ -1,8 +1,8 @@
 import { createStore, combineReducers, applyMiddleware, compose } from 'redux';
 import thunkMiddleware from 'redux-thunk';
 
-import grade from './reducers/grade';
-import enrollment from './reducers/enrollment';
+import grade from './grades/reducers';
+import enrollment from './enrollment/reducers';
 import authReducer from './auth/reducer';
 
 import { commonReducer } from './common/reducer';
@@ -17,7 +17,7 @@ const reducer = combineReducers({
 
 export type ReduxState = ReturnType<typeof reducer>;
 
-export const useReduxSelector: TypedUseSelectorHook<ReduxState> = useSelector
+export const useReduxSelector: TypedUseSelectorHook<ReduxState> = useSelector;
 
 const composeEnhancers =
 	(window && (window as any).__REDUX_DEVTOOLS_EXTENSION_COMPOSE__) || compose;
